Separate patches into groups. Don't fail startup unless essential patches fail
